fix: 내 챌린지 보관함 중간푸시
(cherry picked from commit 93f4d20743b3c244b9a40a0b8c692d002d30d74c)
