+++ src/app/api/user/[user-id]/challenge/route.ts
@@ -24,7 +24,20 @@ export async function GET(
 
   const { data, error } = await supabase
     .from("challenge")
-    .select(`*, user: users (nickname, profile_image_url)`)
+    .select(
+      `
+      *,
+      milestone(
+        *,
+        routine(*),
+        routine_done_daily(*)
+      ),
+      user:users (
+        nickname,
+        profile_image_url
+      )
+    `
+    )
     .eq("user_id", userId)
     // .range(page * limit, page * limit + (limit - 1))
     .order("created_at", { ascending: false })
